Fix db users

+++ DSCMS/Controllers/LayoutsController.cs
@@ -18,13 +18,82 @@ namespace DSCMS.Controllers
 
         public LayoutsController(ApplicationDbContext context)
         {
-            _context = context;    
+            _context = context;
         }
 
         // GET: Layouts
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Layouts.ToListAsync());
+            try
+            {
+                // First, try to fix any NULL values in the database
+                await FixNullLayouts();
+                
+                var layouts = await _context.Layouts.ToListAsync();
+                return View(layouts);
+            }
+            catch (Exception ex)
+            {
+                // If there's an error loading layouts, try to create a default one
+                ViewBag.ErrorMessage = $"Error loading layouts: {ex.Message}. Attempting to create default layout...";
+                
+                try
+                {
+                    await CreateDefaultLayoutIfNone();
+                    var layouts = await _context.Layouts.ToListAsync();
+                    return View(layouts);
+                }
+                catch (Exception innerEx)
+                {
+                    ViewBag.ErrorMessage = $"Failed to create default layout: {innerEx.Message}";
+                    return View(new List<Layout>());
+                }
+            }
+        }
+
+        private async Task FixNullLayouts()
+        {
+            var layouts = await _context.Layouts.ToListAsync();
+            bool hasChanges = false;
+            
+            foreach (var layout in layouts)
+            {
+                if (string.IsNullOrEmpty(layout.Name))
+                {
+                    layout.Name = $"Layout {layout.LayoutId}";
+                    hasChanges = true;
+                }
+                
+                if (string.IsNullOrEmpty(layout.FileLocation))
+                {
+                    layout.FileLocation = "/Views/DSCMS/Layouts/_BootstrapBlog.cshtml";
+                    hasChanges = true;
+                }
+                
+                // FileContents can be null - that's acceptable, so we don't fix it
+            }
+            
+            if (hasChanges)
+            {
+                await _context.SaveChangesAsync();
+            }
+        }
+        
+        private async Task CreateDefaultLayoutIfNone()
+        {
+            var layoutCount = await _context.Layouts.CountAsync();
+            if (layoutCount == 0)
+            {
+                var defaultLayout = new Layout
+                {
+                    Name = "Bootstrap Blog Layout",
+                    FileLocation = "/Views/DSCMS/Layouts/_BootstrapBlog.cshtml", 
+                    FileContents = null // Reference external file, no inline content needed
+                };
+                
+                _context.Layouts.Add(defaultLayout);
+                await _context.SaveChangesAsync();
+            }
         }
 
         // GET: Layouts/Details/5
@@ -35,7 +104,8 @@ namespace DSCMS.Controllers
                 return NotFound();
             }
 
-            var layout = await _context.Layouts.SingleOrDefaultAsync(m => m.LayoutId == id);
+            var layout = await _context.Layouts
+                .FirstOrDefaultAsync(m => m.LayoutId == id);
             if (layout == null)
             {
                 return NotFound();
@@ -51,17 +121,15 @@ namespace DSCMS.Controllers
         }
 
         // POST: Layouts/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("LayoutId,FileContents,FileLocation,Name")] Layout layout)
+        public async Task<IActionResult> Create([Bind("LayoutId,Name,FileLocation,FileContents")] Layout layout)
         {
             if (ModelState.IsValid)
             {
                 _context.Add(layout);
                 await _context.SaveChangesAsync();
-                return RedirectToAction("Index");
+                return RedirectToAction(nameof(Index));
             }
             return View(layout);
         }
@@ -74,7 +142,7 @@ namespace DSCMS.Controllers
                 return NotFound();
             }
 
-            var layout = await _context.Layouts.SingleOrDefaultAsync(m => m.LayoutId == id);
+            var layout = await _context.Layouts.FindAsync(id);
             if (layout == null)
             {
                 return NotFound();
@@ -83,11 +151,9 @@ namespace DSCMS.Controllers
         }
 
         // POST: Layouts/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("LayoutId,FileContents,FileLocation,Name")] Layout layout)
+        public async Task<IActionResult> Edit(int id, [Bind("LayoutId,Name,FileLocation,FileContents")] Layout layout)
         {
             if (id != layout.LayoutId)
             {
@@ -112,7 +178,7 @@ namespace DSCMS.Controllers
                         throw;
                     }
                 }
-                return RedirectToAction("Index");
+                return RedirectToAction(nameof(Index));
             }
             return View(layout);
         }
@@ -125,7 +191,8 @@ namespace DSCMS.Controllers
                 return NotFound();
             }
 
-            var layout = await _context.Layouts.SingleOrDefaultAsync(m => m.LayoutId == id);
+            var layout = await _context.Layouts
+                .FirstOrDefaultAsync(m => m.LayoutId == id);
             if (layout == null)
             {
                 return NotFound();
@@ -139,10 +206,13 @@ namespace DSCMS.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var layout = await _context.Layouts.SingleOrDefaultAsync(m => m.LayoutId == id);
-            _context.Layouts.Remove(layout);
-            await _context.SaveChangesAsync();
-            return RedirectToAction("Index");
+            var layout = await _context.Layouts.FindAsync(id);
+            if (layout != null)
+            {
+                _context.Layouts.Remove(layout);
+                await _context.SaveChangesAsync();
+            }
+            return RedirectToAction(nameof(Index));
         }
 
         private bool LayoutExists(int id)
